backend/pyproject.toml
[project]
name = "couture-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "numpy>=1.24",
    "fpdf2>=2.7",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "httpx>=0.25"]

[tool.setuptools.packages.find]
include = ["app*"]
